requirements.txt
djongo
Django==3.0.2
djangorestframework==3.11.0
redis==3.3.11
mysqlclient==1.4.4
django-filter==2.2.0
Pillow==7.0.0
grequests==0.4.0
aiohttp==3.6.2
drf-yasg==1.17.0
django-cors-headers
gunicorn
django-rest-passwordreset
django-admin-interface
stripe
paypalrestsdk
PyVimeo
channels
channels-redis
icecream
asgiref
aioredis
#asgi_redis